0_common/requirements.txt
pika
google-cloud-storage
google-cloud-pubsub
google-cloud-logging